refactor: move auth.py to src/utils/token_helper.py
remove old db.py by the way.
