--- src/api/utils/auth.py
@@ -1,20 +1,0 @@
-import os
-import jwt
-
-from api.utils.grpc_exception import Unauthenticated
-
-SECRET_KEY = os.getenv("JWT_SECRET_KEY", "default_secret")
-ALGORITHM = "HS512"
-
-def get_token(token: str | bytes | None) -> dict:
-    if token is None:
-        raise Unauthenticated(details="No token provided")
-    try:
-        return jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-    except jwt.ExpiredSignatureError:
-        raise Unauthenticated(details="Token has expired")
-    except jwt.InvalidTokenError:
-        raise Unauthenticated(details="Invalid token")
-
-def generate_token(user_id: str) -> str:
-    return jwt.encode({"sub": user_id}, SECRET_KEY, algorithm=ALGORITHM)
